% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelLoading.R
\name{panels_available}
\alias{panels_available}
\title{List Available Panels}
\usage{
panels_available()
}
\value{
valid panel names character()
}
\description{
List Available Panels
}
\examples{
panels_available()
}
