% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelLoading.R
\name{panels_load}
\alias{panels_load}
\title{Load a Panel}
\usage{
panels_load(panel, verbose = TRUE)
}
\arguments{
\item{panel}{name of a valid microbial panel}

\item{verbose}{controls verbosity of error message (flag)}
}
\value{
dataframe describing microbes in the panel (including ncbi taxonomy ID
}
\description{
Load a Panel
}
\examples{
panels_load('CancerViruses')

}
